// src/app/entities/definicion-producto-interface.ts
export interface DefinicionProductoInterface {
  id?: number;
  id_definicion?: number;
  id_ugp?: number;
  id_accion?: number;
  df1: string;
  df2: string;
  df3: string;
  df4: string;
  df5: string;
  df6: string;
  df7: string;
  df8: string;
  cr1: string;
  cr2: string;
  cr3: string;
  cr4: string;
  cr5: string;
}